// frontend/src/app/routes/organigrama/organigramaRamas_Subramas/types/rama.type.ts
export interface Rama {
  section_id: string;
  sectionName: string;
  sectionDescription?: string;
  sectionGalleryObjectIds: string[];
  id: string;
  nombre: string;
  descripcion?: string;
  icono?: string;
  iconoObjectId?: string; // ID para la imagen en localStorage
  imagenPrincipal?: string; // URL de la imagen principal
  imagenPrincipalObjectId?: string; // ID para la imagen principal en localStorage
  edadMinima: number;
  edadMaxima: number;
  año: number;
  estado: 'activa' | 'inactiva';
  fechaCreacion: string;
  subramas: Subrama[];
}

export interface Subrama {
  subgroup_id: string; 
  subgroupName: string;
  subgroupDescription?: string;
  subgroupGalleryObjectIds?: string[]; // IDs de imágenes de galería en localStorage
  section_id: string; 
  id: string; 
  nombre: string;
  descripcion?: string;
  icono?: string;
  iconoObjectId?: string; // ID para la imagen en localStorage
  imagenPrincipal?: string; // URL de la imagen principal
  imagenPrincipalObjectId?: string; // ID para la imagen principal en localStorage
  ramaId: string;
  lider?: string;
  estado: 'activa' | 'inactiva';
  fechaCreacion: string;
  numeroMiembros: number;
}
// ...
export interface CreateRamaData {
  nombre: string;
  descripcion?: string;
  edadMinima: number;
  edadMaxima: number;
  año: number;
  iconFile?: File;
  galleryFiles?: File[];
}

export interface UpdateRamaData {
  id: string;
  nombre?: string;
  descripcion?: string;
  edadMinima?: number;
  edadMaxima?: number;
  estado?: 'activa' | 'inactiva';
  iconFile?: File;
  galleryFiles?: File[];
}

export interface CreateSubramaData {
  nombre: string;
  descripcion?: string;
  ramaId: string;
  lider?: string;
  galleryFiles?: File[];
}

export interface UpdateSubramaData {
  id: string;
  subgroup_id?: string;
  nombre?: string;
  descripcion?: string;
  lider?: string;
  estado?: 'activa' | 'inactiva';
  ramaId?: string;
}